blockchain, state: refactored state migration related codes
